backend/requirements.txt
autopep8==1.5.7
attrs==20.3.0
backcall==0.2.0
borb
InvoiceGenerator
black
certifi==2020.12.5
chardet==4.0.0
decorator==5.1.1
Django==3.2
django-environ==0.4.5
django-extensions==3.1.3
django-redis==4.12.1
flake8==3.9.1
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent>21.1.2
greenlet>=1.1.0,<2.0
gunicorn==20.1.0
idna>=2.5,<3
isort
ipython>7.30.0
ipython-genutils==0.2.0
jedi==0.18.0
mccabe==0.6.1
mypy==0.812
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
rope==0.19.0
simplegeneric==0.8.1
six==1.15.0
traitlets==5.0.5
typed-ast==1.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
djangorestframework-simplejwt
djangorestframework==3.12.4
drf-nested-routers
